# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.CampaignSessionTie:
    properties:
      briefDescription:
        type: string
      cachedScheduledAt:
        type: string
      cachedTitle:
        type: string
      campaignId:
        type: string
      orderIndex:
        type: integer
      sessionId:
        type: string
    type: object
...
  dtos.SessionApplication:
    properties:
      applicantId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      resolvedAt:
        type: string
      sessionId:
        type: string
      status:
        $ref: '#/definitions/dtos.SessionApplicationStatus'
    type: object
...
  dtos.SessionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dtos.Session'
        type: array
      nextCursor:
        type: string
      users:
        additionalProperties:
          $ref: '#/definitions/dtos.UserBrief'
        type: object
    type: object
...
  internal_handlers.EditTieRequest:
    properties:
      briefDescription:
        type: string
      orderIndex:
        type: integer
    type: object
  internal_handlers.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handlers.TieSessionRequest:
    properties:
      briefDescription:
        type: string
      orderIndex:
        type: integer
      sessionId:
        type: string
    type: object
host: localhost:5137
info:
  contact: {}
  description: Session, game system, campaign and character API for Session Zero
  title: Session Zero Session Service
  version: "1.0"
paths:
...
  /v1/campaigns/{id}/sessions:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CampaignSessionTie'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List sessions tied to a campaign
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session tie data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.TieSessionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Tie a session to a campaign
      tags:
      - campaigns
  /v1/campaigns/{id}/sessions/{sessionId}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Remove a session from a campaign
      tags:
      - campaigns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Tie metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EditTieRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update campaign-session tie metadata
      tags:
      - campaigns
...
  /v1/sessions:
    get:
      parameters:
      - description: Scope
        enum:
        - catalog
        - mastering
        - playing
        in: query
        name: scope
        type: string
      - description: Filter by master user ID
        in: query
        name: masterId
        type: string
      - description: Filter by player user ID
        in: query
        name: playerId
        type: string
      - collectionFormat: csv
        description: Status filter; use 'public' for published/ongoing/completed
        in: query
        items:
          type: string
        name: status
        type: array
      - description: Full-text search
        in: query
        name: search
        type: string
      - description: Format
        enum:
        - online
        - offline
        in: query
        name: format
        type: string
      - description: Type
        enum:
        - oneshot
        - campaign
        in: query
        name: type
        type: string
      - description: City (offline sessions only)
        in: query
        name: city
        type: string
      - collectionFormat: csv
        description: Exclude game system by ID
        in: query
        items:
          type: string
        name: systemExcluded
        type: array
      - collectionFormat: csv
        description: Include only game system by ID
        in: query
        items:
          type: string
        name: systemIncluded
        type: array
      - description: Minimum available seats
        in: query
        name: freeSeats
        type: integer
      - description: Minimum price
        in: query
        name: priceMin
        type: number
      - description: Maximum price
        in: query
        name: priceMax
        type: number
      - description: Start date filter (RFC3339)
        in: query
        name: dateFrom
        type: string
      - description: End date filter (RFC3339)
        in: query
        name: dateTo
        type: string
      - description: Sort field
        enum:
        - scheduled_at
        - created_at
        - price
        - title
        - system
        in: query
        name: sort
        type: string
      - description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.SessionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List sessions
      tags:
      - sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a session
      tags:
      - sessions
...
  /v1/sessions/{id}/applications:
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.SessionApplication'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List applications for a session
      tags:
      - sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.SessionApplication'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Apply to join a session
      tags:
      - sessions
  /v1/sessions/{id}/applications/{applicationId}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Accept or reject an application
      tags:
      - sessions
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"